+++ tests/table/apis/table.entity.test.ts
@@ -1016,4 +1016,237 @@ describe("table Entity APIs test", () => {
       );
     });
   });
+
+  // this covers the following issues
+  // https://github.com/Azure/Azurite/issues/750
+  // https://github.com/Azure/Azurite/issues/733
+  // https://github.com/Azure/Azurite/issues/745
+  it("Operates on batch items with complex row keys, @loki", (done) => {
+    requestOverride.headers = {
+      Prefer: "return-content",
+      accept: "application/json;odata=fullmetadata"
+    };
+    const insertEntity1 = createBasicEntityForTest();
+    insertEntity1.RowKey._ = "8b0a63c8-9542-49d8-9dd2-d7af9fa8790f_0B";
+    const insertEntity2 = createBasicEntityForTest();
+    insertEntity2.RowKey._ = "8b0a63c8-9542-49d8-9dd2-d7af9fa8790f_0C";
+    const insertEntity3 = createBasicEntityForTest();
+    insertEntity3.RowKey._ = "8b0a63c8-9542-49d8-9dd2-d7af9fa8790f_0D";
+    const insertEntity4 = createBasicEntityForTest();
+    insertEntity4.RowKey._ = "8b0a63c8-9542-49d8-9dd2-d7af9fa8790f_0E";
+
+    tableService.insertEntity<TestEntity>(tableName, insertEntity1, () => {
+      tableService.insertEntity<TestEntity>(tableName, insertEntity2, () => {
+        const entityBatch: Azure.TableBatch = new Azure.TableBatch();
+        entityBatch.insertEntity(insertEntity3, { echoContent: true });
+        entityBatch.insertEntity(insertEntity4, {});
+        entityBatch.deleteEntity(insertEntity1);
+        entityBatch.deleteEntity(insertEntity2);
+        tableService.executeBatch(
+          tableName,
+          entityBatch,
+          (batchError, batchResult, batchResponse) => {
+            if (batchError) {
+              assert.ifError(batchError);
+              done();
+            } else {
+              assert.equal(batchResponse.statusCode, 202);
+              tableService.retrieveEntity<TestEntity>(
+                tableName,
+                insertEntity3.PartitionKey._,
+                insertEntity3.RowKey._,
+                (error: any, result, response) => {
+                  assert.strictEqual(
+                    response.statusCode,
+                    200,
+                    "We did not find the 3rd entity!"
+                  );
+                  tableService.retrieveEntity<TestEntity>(
+                    tableName,
+                    insertEntity4.PartitionKey._,
+                    insertEntity4.RowKey._,
+                    (error2: any, result2, response2) => {
+                      assert.strictEqual(
+                        response2.statusCode,
+                        200,
+                        "We did not find the 4th entity!"
+                      );
+                      tableService.retrieveEntity<TestEntity>(
+                        tableName,
+                        insertEntity1.PartitionKey._,
+                        insertEntity1.RowKey._,
+                        (error3: any, result3, response3) => {
+                          assert.strictEqual(
+                            response3.statusCode,
+                            404,
+                            "We did not delete the 1st entity!"
+                          );
+                          tableService.retrieveEntity<TestEntity>(
+                            tableName,
+                            insertEntity2.PartitionKey._,
+                            insertEntity2.RowKey._,
+                            (error4: any, result4, response4) => {
+                              assert.strictEqual(
+                                response4.statusCode,
+                                404,
+                                "We did not delete the 2nd entity!"
+                              );
+                              done();
+                            }
+                          );
+                        }
+                      );
+                    }
+                  );
+                }
+              );
+            }
+          }
+        );
+      });
+    });
+  });
+
+  // this covers https://github.com/Azure/Azurite/issues/741
+  it("Operates on batch items with complex partition keys, @loki", (done) => {
+    requestOverride.headers = {
+      Prefer: "return-content",
+      accept: "application/json;odata=fullmetadata"
+    };
+    const insertEntity1 = createBasicEntityForTest();
+    insertEntity1.PartitionKey._ =
+      "@DurableTask.AzureStorage.Tests.AzureStorageScenarioTests+Orchestrations+AutoStartOrchestration+Responder";
+    const insertEntity2 = createBasicEntityForTest();
+    insertEntity2.PartitionKey._ =
+      "@DurableTask.AzureStorage.Tests.AzureStorageScenarioTests+Orchestrations+AutoStartOrchestration+Responder";
+    const insertEntity3 = createBasicEntityForTest();
+    insertEntity3.PartitionKey._ =
+      "@DurableTask.AzureStorage.Tests.AzureStorageScenarioTests+Orchestrations+AutoStartOrchestration+Responder";
+    const insertEntity4 = createBasicEntityForTest();
+    insertEntity4.PartitionKey._ =
+      "@DurableTask.AzureStorage.Tests.AzureStorageScenarioTests+Orchestrations+AutoStartOrchestration+Responder";
+
+    tableService.insertEntity<TestEntity>(tableName, insertEntity1, () => {
+      tableService.insertEntity<TestEntity>(tableName, insertEntity2, () => {
+        const entityBatch: Azure.TableBatch = new Azure.TableBatch();
+        entityBatch.insertEntity(insertEntity3, { echoContent: true });
+        entityBatch.insertEntity(insertEntity4, {});
+        entityBatch.deleteEntity(insertEntity1);
+        entityBatch.deleteEntity(insertEntity2);
+        tableService.executeBatch(
+          tableName,
+          entityBatch,
+          (batchError, batchResult, batchResponse) => {
+            if (batchError) {
+              assert.ifError(batchError);
+              done();
+            } else {
+              assert.equal(batchResponse.statusCode, 202);
+              tableService.retrieveEntity<TestEntity>(
+                tableName,
+                insertEntity3.PartitionKey._,
+                insertEntity3.RowKey._,
+                (error: any, result, response) => {
+                  assert.strictEqual(
+                    response.statusCode,
+                    200,
+                    "We did not find the 3rd entity!"
+                  );
+                  tableService.retrieveEntity<TestEntity>(
+                    tableName,
+                    insertEntity4.PartitionKey._,
+                    insertEntity4.RowKey._,
+                    (error2: any, result2, response2) => {
+                      assert.strictEqual(
+                        response2.statusCode,
+                        200,
+                        "We did not find the 4th entity!"
+                      );
+                      tableService.retrieveEntity<TestEntity>(
+                        tableName,
+                        insertEntity1.PartitionKey._,
+                        insertEntity1.RowKey._,
+                        (error3: any, result3, response3) => {
+                          assert.strictEqual(
+                            response3.statusCode,
+                            404,
+                            "We did not delete the 1st entity!"
+                          );
+                          tableService.retrieveEntity<TestEntity>(
+                            tableName,
+                            insertEntity2.PartitionKey._,
+                            insertEntity2.RowKey._,
+                            (error4: any, result4, response4) => {
+                              assert.strictEqual(
+                                response4.statusCode,
+                                404,
+                                "We did not delete the 2nd entity!"
+                              );
+                              done();
+                            }
+                          );
+                        }
+                      );
+                    }
+                  );
+                }
+              );
+            }
+          }
+        );
+      });
+    });
+  });
+
+  it("Ensure Valid Etag format from Batch, @loki", (done) => {
+    requestOverride.headers = {
+      Prefer: "return-content",
+      accept: "application/json;odata=fullmetadata"
+    };
+    const batchEntity1 = createBasicEntityForTest();
+
+    const entityBatch: Azure.TableBatch = new Azure.TableBatch();
+    entityBatch.addOperation("INSERT", batchEntity1, { echoContent: true });
+    batchEntity1.myValue._ = "value2";
+    entityBatch.mergeEntity(batchEntity1);
+
+    tableService.executeBatch(
+      tableName,
+      entityBatch,
+      (updateError, updateResult, updateResponse) => {
+        if (updateError) {
+          assert.ifError(updateError);
+          done();
+        } else {
+          assert.equal(updateResponse.statusCode, 202);
+          tableService.retrieveEntity<TestEntity>(
+            tableName,
+            batchEntity1.PartitionKey._,
+            batchEntity1.RowKey._,
+            (error, result, response) => {
+              if (error) {
+                assert.ifError(error);
+              } else if (result) {
+                const entity: TestEntity = result;
+                assert.equal(entity.myValue._, batchEntity1.myValue._);
+
+                if (response !== null) {
+                  const body: any = response?.body;
+                  assert.notStrictEqual(body, null, "response empty");
+                  if (body != null) {
+                    assert.strictEqual(
+                      body["odata.etag"].match(/(%3A)/).length,
+                      2,
+                      "did not find the expected number of escaped sequences"
+                    );
+                  }
+                }
+              }
+              done();
+            }
+          );
+        }
+      }
+    );
+  });
 });